feat: move process interaction into a different module
