+++ bin/main.ts
@@ -1,10 +1,10 @@
-import { InteractionType } from 'discord-api-types/v10';
 import { Result } from 'oxide.ts';
 import { getDiscordClient } from '../src/clients';
 import { getConfigs } from '../src/config';
 import { commands as contextMenuCommandList } from '../src/context-menu-commands';
 import { deployGlobalCommands } from '../src/deploy-command';
 import { commands as slashCommandList } from '../src/slash-commands';
+import { processInteraction } from '../src/utils/interaction-processor';
 import { loadEnv } from '../src/utils/load-env';
 import { logger } from '../src/utils/logger';
 import { processMessage } from '../src/utils/message-processor';
@@ -42,33 +42,7 @@ const main = async () => {
   });
 
   client.on('interactionCreate', async (interaction) => {
-    try {
-      const isCommand = interaction.isChatInputCommand();
-      if (isCommand) {
-        const { commandName } = interaction;
-        logger.info(`[main]: RECEIVED COMMAND. COMMAND: ${commandName}`);
-        const command = slashCommandList.find((cmd) => cmd.data.name === commandName);
-        return await command?.execute(interaction);
-      }
-
-      const isContextMenuCommand = interaction.isContextMenuCommand();
-      if (isContextMenuCommand) {
-        const { commandName } = interaction;
-        logger.info(`[main]: RECEIVED CONTEXT MENU COMMAND. COMMAND: ${commandName}`);
-        const command = contextMenuCommandList.find((cmd) => cmd.data.name === commandName);
-        return await command?.execute(interaction);
-      }
-
-      const isAutocomplete = interaction.type === InteractionType.ApplicationCommandAutocomplete;
-      if (isAutocomplete) {
-        const { commandName } = interaction;
-        logger.info(`[main]: RECEIVED AUTOCOMPLETE. COMMAND: ${commandName}`);
-        const command = slashCommandList.find((cmd) => cmd.data.name === commandName);
-        return await command?.autocomplete?.(interaction);
-      }
-    } catch (error) {
-      logger.error(`[main]: ERROR HANDLING INTERACTION, ERROR: ${error}`);
-    }
+    return processInteraction(interaction);
   });
 };
